+++ infrastructure/src/main/java/com/arthurlamberti/cdb/infrastructure/wallet/models/GetWalletResponse.java
@@ -2,7 +2,7 @@ package com.arthurlamberti.cdb.infrastructure.wallet.models;
 
 public record GetWalletResponse (
         String id,
-        Double amount,
+        Integer amount,
         String customerId
 ){
 }
